Conflicts:
	bmc-api/src/main/java/com/ai/baas/bmc/api/failedbillmaintain/interfaces/IFailedBillMaintainSV.java
	bmc-api/src/main/java/com/ai/baas/bmc/api/failedbillmaintain/params/FailedBill.java
	bmc-api/src/main/java/com/ai/baas/bmc/api/failedbillmaintain/params/FailedBillCriteria.java
	bmc-api/src/main/java/com/ai/baas/bmc/api/failedbillmaintain/params/FailedBillParam.java
	bmc-core/build.gradle
	bmc-core/src/main/java/com/ai/baas/bmc/api/failedbillmaintain/iml/FailedBillMaintainSVImpl.java
	bmc-core/src/main/java/com/ai/baas/bmc/business/impl/FailedBillMaintainBusiImpl.java
	bmc-core/src/main/java/com/ai/baas/bmc/business/interfaces/IFailedBillMaintainBusi.java
	bmc-core/src/main/java/com/ai/baas/bmc/dao/interfaces/BmcRecordFmtMapper.java
	bmc-core/src/main/java/com/ai/baas/bmc/util/MDSUtil.java
	bmc-core/src/main/resources/mds.properties
	build.gradle

+++ bmc-core/src/main/java/com/ai/baas/bmc/dao/interfaces/BmcBasedataCodeMapper.java
@@ -10,7 +10,7 @@ public interface BmcBasedataCodeMapper {
 
     int deleteByExample(BmcBasedataCodeCriteria example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(BmcBasedataCode record);
 
@@ -18,7 +18,7 @@ public interface BmcBasedataCodeMapper {
 
     List<BmcBasedataCode> selectByExample(BmcBasedataCodeCriteria example);
 
-    BmcBasedataCode selectByPrimaryKey(Integer id);
+    BmcBasedataCode selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") BmcBasedataCode record, @Param("example") BmcBasedataCodeCriteria example);
